# app_next/src/lib/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tank Battle API",
    "description": "API for Tank Battle - a multiplayer tank game.\n\n**Authentication:**\n- **Public API**: Use `X-API-Key` header with a valid API key (generate via `POST /api/me/api-keys` with session auth).\n- **Session API**: Requires a valid NextAuth session (cookie-based).",
    "version": "1.1.0",
    "contact": {
      "name": "Tank Battle API Support"
    }
  },
  "servers": [
    {
      "url": "https://localhost:8443",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key generated from your profile. Format: `pk_<uuid>`. Generate keys via `POST /api/me/api-keys` (requires session auth)."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie. Obtained by signing in via the web application."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Invalid API key" }
        }
      },
...
      "ProfileDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "username": { "type": "string", "example": "TankMaster" },
          "email": { "type": "string", "format": "email", "example": "tank@battle.com" },
          "tankColor": { "type": "string", "example": "#ff0000" },
          "tankLevel": { "type": "integer", "example": 12 },
          "xp": { "type": "integer", "example": 5400 },
          "wins": { "type": "integer", "example": 42 },
          "gamesPlayed": { "type": "integer", "example": 100 },
          "isOnline": { "type": "boolean", "example": true },
          "lastSeen": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "gamesAsPlayer": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GameHistory" }
          }
        }
      },
...
      "GameHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "opponentId": { "type": "integer", "example": 2 },
          "winnerId": { "type": "integer", "example": 1 },
          "playerScore": { "type": "integer", "example": 10 },
          "opponentScore": { "type": "integer", "example": 7 },
          "duration": { "type": "integer", "example": 180, "description": "Duration in seconds" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ProfileUpdateBody": {
        "type": "object",
        "properties": {
          "tankColor": { "type": "string", "example": "#00ff00", "description": "Hex color code" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": ["Me"],
        "summary": "Get current user",
        "description": "Retrieve the authenticated user's profile.",
        "security": [{ "SessionAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/ProfileDetail" } } } } }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Me"],
        "summary": "Update current user",
        "description": "Update the authenticated user's tank color.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileUpdateBody" } } }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/ProfileDetail" } } } } }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/me/online": {
      "get": {
        "tags": ["Me"],
        "summary": "Get online status",
        "description": "Get the authenticated user's online status.",
        "security": [{ "SessionAuth": [] }],
        "responses": {
          "200": {
            "description": "Online status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "isOnline": { "type": "boolean" },
                    "lastSeen": { "type": "string", "format": "date-time", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Me"],
        "summary": "Update online status",
        "description": "Set the authenticated user's online/offline status.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isOnline"],
                "properties": {
                  "isOnline": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "id": { "type": "integer" }, "isOnline": { "type": "boolean" }, "lastSeen": { "type": "string", "format": "date-time" } } } } }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
...
}